+++ tests/testthat/test-quantization.R
@@ -8,13 +8,17 @@ offset = 500
 LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
 las = readLAS(LASfile, select = "xyz")
 
+test_that("is.quantized works", {
+  expect_true(!is.quantized(x, scale, offset))
+})
+
 test_that("Quantization works", {
 
-  lidR:::fast_quantization(x, scale, offset)
+  quantize(x, scale, offset)
 
   expect_equal(x, round(y, 3), tolerance = 1e-10)
-  expect_equal(lidR:::fast_countunquantized(x, scale, offset), 0L)
-  expect_equal(lidR:::fast_countunquantized(y, scale, offset), 100L)
+  expect_equal(count_not_quantized(x, scale, offset), 0L)
+  expect_equal(count_not_quantized(y, scale, offset), 100L)
 })
 
 test_that("Quantization works on las data", {
@@ -26,16 +30,17 @@ test_that("Quantization works on las data", {
   zscale = las@header@PHB$`Z scale factor`
   zoffset = las@header@PHB$`Z offset`
 
-  expect_equal(lidR:::fast_countunquantized(las$X, xscale, xoffset), 0L)
-  expect_equal(lidR:::fast_countunquantized(las$Y, yscale, yoffset), 0L)
-  expect_equal(lidR:::fast_countunquantized(las$Z, zscale, zoffset), 0L)
+  expect_equal(count_not_quantized(las$X, xscale, xoffset), 0L)
+  expect_equal(count_not_quantized(las$Y, yscale, yoffset), 0L)
+  expect_equal(count_not_quantized(las$Z, zscale, zoffset), 0L)
 
   x = las@data[["X"]][1]
   las@data[["X"]][1] = x + 0.0001
 
-  expect_equal(lidR:::fast_countunquantized(las$X, xscale, xoffset), 1L)
+  expect_equal(count_not_quantized(las$X, xscale, xoffset), 1L)
 
-  lidR:::fast_quantization(las@data[["X"]], xscale, xoffset)
+  las_quantize(las)
 
   expect_equal(las@data[["X"]][1], x,  tolerance = 0)
 })
+